+++ yamlwav/__init__.py
@@ -4,4 +4,4 @@ from .decoder import decode
 from .encoder import encode, encode_dict
 
 __all__ = ["encode", "encode_dict", "decode", "WavConfig"]
-__version__ = "1.0.1"
+__version__ = "2.0.0"